dom/tests/mochitest/chrome/test_elements_proto.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>

<window title="Mozilla Bug 861493"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <body xmlns="http://www.w3.org/1999/xhtml">
  </body>

  <script type="application/javascript">
  <![CDATA[
    const XUL_NS = "http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul";

    SimpleTest.waitForExplicitFinish();

    addLoadEvent(() => {
      ok(XULTextElement.isInstance(document.createElementNS(XUL_NS, "label")),
        `<label> should be of XULTextElement type`);

      ok(XULTextElement.isInstance(document.createElementNS(XUL_NS, "description")),
        `<description> should be of XULTextElement type`);

      SimpleTest.finish();
    });
  ]]>
  </script>
</window>
